+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 #### 1.1.0 (UNRELEASED)
 
-* Fixed `write` to cap SVG width, height and viewBox to a minimum of 1px 
+* Fixed `write` to cap SVG width and height to a minimum of 1px (#102)
 * Fixed grouping of `stat` command in `vpype --help`
 * Bump svgelements from 1.3.2 to 1.3.4 (#101)