+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Invisible SVG elements are now discarded (#103)
 * Fixed `write` to cap SVG width and height to a minimum of 1px (#102)
 * Fixed grouping of `stat` command in `vpype --help`
+* API: added vpype_cli.execute() to execute a vpype pipeline from python (#104)
 * Bump svgelements from 1.3.2 to 1.3.4 (#101)