+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 #### 1.1.0 (UNRELEASED)
 
+New features and improvements:
+* Added `snap` command (#110)
 * Invisible SVG elements are now discarded (#103)
+
+API changes:
+* Added vpype_cli.execute() to execute a vpype pipeline from python (#104)
+
+Bug fixes:
 * Fixed `write` to cap SVG width and height to a minimum of 1px (#102)
 * Fixed grouping of `stat` command in `vpype --help`
-* API: added vpype_cli.execute() to execute a vpype pipeline from python (#104)
-* Bump svgelements from 1.3.2 to 1.3.4 (#101)
 
 
 #### 1.0.0 (2020-11-29)